add scripts for running tests on a cluster like Hotfoot
The usual "make -j" approach to testing in parallel is unsuited
for Hotfoot because it limits nodes to 12 processors. In addition,
with 12 processors, about 12 GB of RAM would be required, which
is only supported by a smaller number of servers that may not be
available at any given time.

Instead, these scripts are intended to be called via a "job array",
which repeatedly calls the script with an incrementing counter. A
job array can be executed with essentially an unlimited number of
nodes each with only 1 processor. By default, most of the nodes
on the Hotfoot cluster are limited to 2 GB of RAM per node, so we
need to ensure that each test is consistent with that or else it
will fail.